Extracting the search card view to its own class to hopefully fix bugs around nested scrolling
